Create `rustc_ty_library`
